+++ src/header/book.h
@@ -4,7 +4,7 @@
 #define title_size 50
 
 struct book{
-	int isbn_number;
+	long int isbn_number;
 	int serial_number;
 	char *author;
 	char *book_title;
@@ -21,5 +21,5 @@ void Delete_book();
 void Update_book();
 void display_book();
 void search_book(char*);
-void search_book_isbn(int);
+void search_book_isbn(long int);
